# Speed up filter accesses like {% if a.b.c.d %}.