--- a/test/templates/any.html
+++ b/test/templates/any.html
@@ -1,0 +1,3 @@
+{% for p in products %}
+    {% if any p.a p.b p.c p.d p.e p.f p.g %}{% endif %}
+{% endfor %}