--- a/test/templates/numerics.html
+++ b/test/templates/numerics.html
@@ -1,0 +1,4 @@
+{% for p in ps %}
+    {% if p.a > p.b %}
+    {% endif %}
+{% endfor %}