Helps a lot with https://github.com/yogthos/Selmer/issues/245, altough I will keep digging further.

I've added an if-bench benchmark, and taken it down from ~145ms to ~30ms.

The main culpit for the slowness was fix-accessor, which was using exception handling
to parse integers. Exceptions are expensive, at least when you use them
a couple of thusand times:

```
((time (dotimes [_ 100000] (try (Long/valueOf "nan")
                               (catch NumberFormatException _))))
"Elapsed time: 219.507474 msecs"

(time (dotimes [_ 100000] (when (re-matches #"\d+" "nan") (Long/valueOf "nan"))))
"Elapsed time: 11.263704 msecs"
```

## Performance of the if-bench benchmark:

#### before:

```
BENCH: Many . acceses in an if clause
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 145.789759 ms
    Execution time std-deviation : 3.397597 ms
   Execution time lower quantile : 140.769053 ms ( 2.5%)
   Execution time upper quantile : 149.232308 ms (97.5%)
                   Overhead used : 7.470498 ns
```

#### after:

```
BENCH: Many . acceses in an if clause
Evaluation count : 30 in 6 samples of 5 calls.
             Execution time mean : 27.158986 ms
    Execution time std-deviation : 7.895735 ms
   Execution time lower quantile : 21.507726 ms ( 2.5%)
   Execution time upper quantile : 38.751057 ms (97.5%)
                   Overhead used : 7.470498 ns
```

I suspect that it could be made even faster but I'll need to do more digging, and this is a big improvement in itself. 
The other big time spender is add-node it seems, but I haven't looked into it further.