@yogthos 2021-01-22
Ah makes sense, and luckily a straight forward fix for a big performance boost. I can keep the pr open for a bit if you want to dig further, or I can do a release with the fix right away if you need the new artifact on Clojars.