@rutchkiwi 2021-01-22
Thanks! Up to you how you wanna do it - if it's easier for you let's keep this PR open and I will hopefully be able to find more improvements to add to it.