@yogthos 2021-01-22
Sounds like a plan, just ping me when you want to do the release. And thanks for spearheading this! :)