@rutchkiwi 2021-01-22
Sure will do. And no worries, just glad to be able to contribute! :)
On 22 Jan 2021, 22:16 +0100, Dmitri Sotnikov <notifications@github.com>, wrote:
> Sounds like a plan, just ping me when you want to do the release. And thanks for spearheading this! :)
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub, or unsubscribe.