@rutchkiwi 2021-01-25
Trying this out in my actual project and indeed it takes the rendering time down by 75% or so. But I have two questions that I was hoping you could help me with:

1. Since the change above is about parsing if statement conditions, why does it even do anything when the template cache is on? I thought the template cache stored some kind of AST like structure which had everything parsed and ready? (and thus improvements to parsing shouldn't affect things once it's in the cache). Or does this pre-parsing not go all the way in these cases?

2. What does the add-node fn do, conceptually? I'm feeling a bit lost on the big picture which makes it a bit tricky to figure out whats going on.

I'd really appreciate your help @yogthos  :)