@yogthos 2021-01-25
Good to hear. Cache only applies to static elements in the template, but anything with dynamic content has to be recomputed for each render. When the template file is parsed then static content is parsed into strings, and any tags that work with the context map are turned into functions. The resulting template that gets cached is a vector containing a combination of strings and functions [here](https://github.com/yogthos/Selmer/blob/deed36e2d8d67d5cd4f2cc226e59ddf2edb9252b/src/selmer/parser.clj#L288). Then when render is called it runs through the nodes in the vector [here](https://github.com/yogthos/Selmer/blob/deed36e2d8d67d5cd4f2cc226e59ddf2edb9252b/src/selmer/parser.clj#L100) to generate the resulting text. The `add-node` function simply appends a node which is either a string or a function to the vector representing the template.