@rutchkiwi 2021-01-25
Okay thanks. That makes it more clear. I hadn't really figured out how the caching works but now I see that it's done using closures which is cool. I managed to speed things up further but I feel that there is still more to be done so will keep looking.