@rutchkiwi 2021-01-26
Sorry about the many questions, but is there an easy way to reload changes in the `tags` NS, like changing the `if-any-all-fn` function? Right now I have to resort to restarting my repl after every code change. I'm not sure exactly why reloading that file itself doesn't do it.