@yogthos 2021-01-26
Oh yeah, the easiest option is to use tools.namespace to refresh. I just checked in an update on master. If you merge it in you should be able to do `(repl/refresh)` from the REPL to reload the namespaces.