@rutchkiwi 2021-01-26
Oh wow. That makes it a lot easier. Thanks!