@rutchkiwi 2021-01-26
Okay looks like I've reached the limits of my optimization skills now! Can't really stop any more clear hotspots or easy fixes.

For comparison, the template rendering that powers [medino.com's category page](https://www.medino.com/category/supplements?up-to-page=50 ) used to run in ~850ms, but with these changes its now down to ~150ms. (so about 5x faster).

I'm not sure if that's the limit, 150ms feels pretty long but then again it is a lot of template to render! Perhaps I'll try and revisit at some later point.

Let me know if you spot something funky in the code, if not a release would be much appreciated. :)