@yogthos 2021-01-26
Definitely looks like a solid improvement, and if I get some time I'll take a look to see if some more optimizations are possible. The 5x improvement is definitely nice though. :)