@rutchkiwi 2021-01-26
Thanks a lot! And if you find any more things to speed up, or any ideas that I can look into it'd be much appreciated :)